refactor: re-write controllers
