+++ ts/server/src/app.controller.ts
@@ -2,9 +2,7 @@ import { Controller, Post, Body, HttpCode, UseFilters } from '@nestjs/common';
 import { AppService } from './app.service';
 import { Resource } from 'fhir/r4b';
 import * as fhirpath_r4_model from 'fhirpath/fhir-context/r4';
-import * as fhirpath from 'fhirpath';
 import { FPMLValidationErrorFilter } from './app.filters';
-import { FPOptions } from './utils/extract';
 class Template {
     context: Record<string, Resource> | Resource;
     template: object;
@@ -27,39 +25,11 @@ export class AppController {
     resolveTemplateR4(@Body() body: Template): object {
         const { context, template, strict = false } = body;
 
-        const options: FPOptions = {
-            userInvocationTable: {
-                answers: {
-                    fn: (inputs, linkId: string) => {
-                        return fhirpath.evaluate(
-                            inputs,
-                            `repeat(item).where(linkId='${linkId}').answer.value`,
-                            null,
-                            fhirpath_r4_model,
-                            null,
-                        );
-                    },
-                    arity: { 0: [], 1: ['String'] },
-                },
-            },
-        };
-        if (containsQuestionnaireResponse(context)) {
-            return this.appService.resolveTemplate(
-                context.QuestionnaireResponse,
-                template,
-                context,
-                fhirpath_r4_model,
-                options,
-                strict,
-            );
-        }
-
         return this.appService.resolveTemplate(
-            context,
+            containsQuestionnaireResponse(context) ? context.QuestionnaireResponse : context,
             template,
             context,
             fhirpath_r4_model,
-            options,
             strict,
         );
     }
@@ -69,31 +39,12 @@ export class AppController {
     resolveTemplateAidbox(@Body() body: Template): object {
         const { context, template, strict = false } = body;
 
-        const options: FPOptions = {
-            userInvocationTable: {
-                answers: {
-                    fn: (inputs, linkId: string) => {
-                        return fhirpath.evaluate(
-                            inputs,
-                            `repeat(item).where(linkId='${linkId}').answer.value.children()`,
-                            null,
-                        );
-                    },
-                    arity: { 0: [], 1: ['String'] },
-                },
-            },
-        };
-        if (containsQuestionnaireResponse(context)) {
-            return this.appService.resolveTemplate(
-                context.QuestionnaireResponse,
-                template,
-                context,
-                null,
-                options,
-                strict,
-            );
-        }
-
-        return this.appService.resolveTemplate(context, template, context, null, options, strict);
+        return this.appService.resolveTemplate(
+            containsQuestionnaireResponse(context) ? context.QuestionnaireResponse : context,
+            template,
+            context,
+            null,
+            strict,
+        );
     }
 }
